Code improvements suggested by SonarQube.
- correct synchronization (although some still ignored because not necessary and we want to avoid many synchronized calls per word indexed)
- preserving interrupted status when catching InterruptedException
- cast to long to show intentionality
- throw NoSuchElementException from Iterator.next()
- use replace instead of replaceAll if not a regex
- don't throw from a finally block
- re-use (Secure)Random objects

+++ engine/src/main/java/nl/inl/blacklab/search/results/hits/fetch/HitPublisherSpans.java
@@ -335,7 +335,7 @@ public class HitPublisherSpans implements HitPublisher {
                             if (hitsCounted > hitsCountedPrev) {
                                 subscribers.counted(hitsCounted - hitsCountedPrev, docsCounted - docsCountedPrev);
                                 hitsStats.add(0, hitsCounted - hitsCountedPrev);
-                                docsStats.add(0, docsCounted - docsCountedPrev);
+                                docsStats.add(0, (long)docsCounted - docsCountedPrev);
                                 hitsCountedPrev = hitsCounted;
                                 docsCountedPrev = docsCounted;
                             }
